move execution logic to lib
